refactoring with navigation added
